Instrument reads of variable and object properties
Variable references (represented in the AST via the class VariableProxy)
are rewritten like |g => ER_read("g", g)|. This is done only for the
properties of the global object - references to function scope, block
scope, parameter, etc. variables are not modified.

Read of object properties is rewritten like

  obj.key  => function() { let $o = obj, $k = key;
                           return ER_readProp($o, $k, $o.$k); }()

The anonymous function is needed so we can ensure |obj| and |key| are
evaluated exactly once and the resulting construct is still an
Expression. In the common case of |key| being a literal, the last
parameter of |ER_read| looks like |$o.key|.

The rewriting itself proceeds in two phases: before and after name
resolution and variable allocation (done by |Scope::Analyze|. The object
properties access is rewritten before scope analysis. Then the scope
analysis is run, which determines which variables are global and also
allocates local variables, including the ones introduced by the
preceding rewrite. Then the second rewriting pass is run, to modify the
global variable reads, now that we know, which are the global variables.