Ensure the AST nodes have unique identifiers within a function
This is a requirement of the optimizing compiler. Assignment of the
unique ids is interleaved with parsing and id ranges for each function
are maintained via the FunctionState (cf. SetupFunctionState,
TearDownFunctionState).  When the ER rewriter needs to create new AST
nodes, it must ensure they do not conflict with existing ones. We
achieve this by recording the maximum id for each function literal and
then adjusting the id allocation via the AstNodeIdAllocationScope class.