Tweak source positions, set during instrumentation
Be conservative and do not set fictitious source positions, for the
cases, where a source position is considered unnecessary.  Especially
make sure that the calls to instrumentation function literals do not
fall withing the scope of the function literal itself, as it confused
the debugger and leads to wrong scope determination and crashes.