Ensure the "next" part of "for" statements have a source position
... by copying it from the expression into the containing expression
statement. This change causes a source position to be emitted by the
full code generator just before the code for the "next" expression.
Missing a source position in that point may cause a break location's
source position to be erroneously deduced to be inside the for loop
body, which in turn may cause a wrong (inactive) scope to be declared
current and ultimately result in assertion failure or worse.